# docs/swagger.yaml
swagger: "2.0"
info:
  title: Estate Agency API
  description: 'Введите токен в формате: "Bearer <token>" (включая слово Bearer и пробел)'
  version: 0.0.1
host: localhost:8080
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
security:
  - BearerAuth: []
paths:
...
  /properties:
    get:
      summary: List properties
      tags: [properties]
      description: Returns paginated list of properties. Requires authentication.
      parameters:
        - name: limit
          in: query
          type: integer
          example: 10
        - name: offset
          in: query
          type: integer
          example: 0
        - name: ids
          in: query
          type: string
          example: "1,2,3"
        - name: type_ids
          in: query
          type: string
          example: "2,3"
        - name: type_id
          in: query
          type: integer
          example: 2
        - name: transaction_type
          in: query
          type: string
          example: "sale"
        - name: city
          in: query
          type: string
          example: "New York"
        - name: property_status
          in: query
          type: string
          example: "available"
        - name: created_by
          in: query
          type: integer
          example: 1
        - name: min_price
          in: query
          type: number
          example: 100000
        - name: max_price
          in: query
          type: number
          example: 500000
        - name: min_area
          in: query
          type: number
          example: 50
        - name: max_area
          in: query
          type: number
          example: 200
        - name: search
          in: query
          type: string
          example: "spacious"
        - name: latitude
          in: query
          type: number
          example: 40.7128
        - name: longitude
          in: query
          type: number
          example: -74.006
        - name: radius_km
          in: query
          type: number
          example: 5
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/propertyhandler.ListPropertiesResponseDoc'
          examples:
            application/json:
              properties:
                - id: 1
                  title: "Luxury Apartment"
                  property_description: "A spacious 3-bedroom apartment."
                  type_id: 2
                  transaction_type: "sale"
                  price: 350000
                  area: 120
                  property_address: "123 Main Street"
                  latitude: 40.7128
                  longitude: -74.006
                  city: "New York"
                  property_status: "available"
                  images:
                    - filename: "photo1.jpg"
                      data: "base64encodedstring=="
                  image:
                    filename: "photo1.jpg"
                    data: "base64encodedstring=="
                  created_at: "2025-11-29T12:00:00Z"
                  updated_at: "2025-11-29T12:00:00Z"
              total: 1
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      summary: Create property
      tags: [properties]
      description: |
        Create a new property. Requires admin role.

        Адрес (`property_address`) должен быть корректным и непустым — при создании сервиса будут получены координаты (``latitude``, ``longitude``) с помощью Geoapify Geocoding API.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/propertyhandler.PropertyCreateDoc'
      responses:
        "201":
          description: Создано
          schema:
            $ref: '#/definitions/propertyhandler.PropertyDTODoc'
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "409":
          description: Конфликт
        "422":
          description: Unprocessable Entity — геокодирование не вернуло координаты
          schema:
            $ref: '#/definitions/GeocodingErrorResponse'
        "502":
          description: Bad Gateway — ошибка внешнего геосервиса
          schema:
            $ref: '#/definitions/GeocodingErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ErrorResponse'
  /properties/{id}:
    get:
      summary: Get property
      tags: [properties]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/propertyhandler.PropertyDTODoc'
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "404":
          description: Не найдено
        "500":
          description: Внутренняя ошибка сервера
    patch:
      summary: Partially update property
      tags: [properties]
      description: |
        Partially update an existing property. Only provided fields will be changed. Requires admin role.

        Example request (partial body):

        {
          "title": "Cozy studio near park",
          "price": 120000,
          "property_description": "Renovated studio, new windows."
        }
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/propertyhandler.UpdatePropertyDoc'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/propertyhandler.PropertyDTODoc'
          examples:
            application/json:
              id: 1
              title: "Cozy studio near park"
              property_description: "Renovated studio, new windows."
              type_id: 2
              transaction_type: "sale"
              price: 120000
              area: 30
              property_address: "456 Park Ave"
              latitude: 40.7128
              longitude: -74.006
              city: "New York"
              property_status: "available"
              images:
                - filename: "photo1.jpg"
                  data: "base64encodedstring=="
              image:
                filename: "photo1.jpg"
                data: "base64encodedstring=="
              created_at: "2025-11-29T12:00:00Z"
              updated_at: "2025-11-29T12:00:00Z"
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "404":
          description: Не найдено
        "422":
          description: Unprocessable Entity — геокодирование не вернуло координаты
          schema:
            $ref: '#/definitions/GeocodingErrorResponse'
        "502":
          description: Bad Gateway — ошибка внешнего геосервиса
          schema:
            $ref: '#/definitions/GeocodingErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
    delete:
      summary: Delete property
      tags: [properties]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              deleted_id:
                type: integer
                example: 1
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "404":
          description: Не найдено
        "500":
          description: Внутренняя ошибка сервера
  /properties/{id}/favorites:
    post:
      summary: Toggle favorite
      description: >-
        Переключает состояние избранного для указанного объявления от лица
        текущего аутентифицированного пользователя. Если объявление добавлено в
        избранное — возвращает созданный объект свойства (PropertyDTO) со
        всеми полями, включая `images` (список изображений) и `image` (первая картинка).
        Если запись удалена — возвращает 204 No Content.
      tags: [properties]
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
      responses:
        "201":
          description: Создано
          schema:
            $ref: '#/definitions/propertyhandler.PropertyDTODoc'
          examples:
            application/json:
              id: 1
              title: "Luxury Apartment"
              property_description: "A spacious 3-bedroom apartment."
              type_id: 2
              transaction_type: "sale"
              price: 350000
              area: 120
              property_address: "123 Main Street"
              latitude: 40.7128
              longitude: -74.006
              city: "New York"
              property_status: "available"
              images:
                - filename: "photo1.jpg"
                  data: "base64encodedstring=="
                - filename: "photo2.jpg"
                  data: "base64encodedstring2=="
              image:
                filename: "photo1.jpg"
                data: "base64encodedstring=="
              created_at: "2025-11-29T12:00:00Z"
              updated_at: "2025-11-29T12:00:00Z"
        "204":
          description: Удалено
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "404":
          description: Не найдено
        "500":
          description: Внутренняя ошибка сервера
  /properties/{id}/images:
    
    put:
      summary: Replace images
      tags: [properties]
      consumes:
        - multipart/form-data
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          example: 1
        - name: files
          in: formData
          description: |
            New images to replace existing ones. Передавайте несколько файлов с полем `files` (multipart/form-data).
            Для лучшей совместимости используем массив файлов: `type: array` с `items: { type: file }` и `collectionFormat: multi`.
            Если ваш клиент/генератор не поддерживает массив файлов - используйте несколько полей `files`.
          required: true
          type: array
          items:
            type: file
          collectionFormat: multi
          x-files-multiple: true
      responses:
        "204":
          description: Нет содержимого
        "400":
          description: Некорректный запрос
        "401":
          description: Требуется авторизация
        "403":
          description: Доступ запрещён
        "500":
          description: Внутренняя ошибка сервера
        "501":
          description: Не реализовано

  # -------------------- Property Types --------------------
...
definitions:
  propertyhandler.PropertyDTODoc:
    type: object
    properties:
      id: { type: integer, example: 1 }
      title: { type: string, example: "Luxury Apartment" }
      property_description: { type: string, example: "A spacious 3-bedroom apartment." }
      type_id: { type: integer, example: 2 }
      transaction_type: { type: string, example: "sale" }
      price: { type: number, example: 350000 }
      area: { type: number, example: 120 }
      property_address: { type: string, example: "123 Main Street", description: "Должен быть корректным, непустым адрес. Координаты (latitude, longitude) вычисляются с помощью Geoapify Geocoding API при создании/обновлении." }
      latitude: { type: number, example: 40.7128 }
      longitude: { type: number, example: -74.006 }
      city: { type: string, example: "New York" }
      property_status: { type: string, example: "available" }
      is_favorited: { type: boolean, example: false }
      created_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
      updated_at: { type: string, format: date-time, example: "2025-11-29T12:00:00Z" }
      images:
        type: array
        items:
          $ref: '#/definitions/imagedto.ImageFileDoc'
      image:
        $ref: '#/definitions/imagedto.ImageFileDoc'
  propertyhandler.PropertyCreateDoc:
    type: object
    required:
      - title
      - property_description
      - price
      - area
      - city
      - property_address
      - transaction_type
      - type_id
    properties:
      title: { type: string, example: "Luxury Apartment" }
      property_description: { type: string, example: "A spacious 3-bedroom apartment." }
      type_id: { type: integer, example: 2 }
      transaction_type: { type: string, example: "sale" }
      price: { type: number, example: 350000 }
      area: { type: number, example: 120 }
      property_address: { type: string, example: "123 Main Street", description: "Должен быть корректным, непустым адрес. Координаты (latitude, longitude) вычисляются с помощью Geoapify Geocoding API при создании/обновлении." }
      city: { type: string, example: "New York" }
  propertyhandler.UpdatePropertyDoc:
    type: object
    properties:
      title: { type: string, example: "Luxury Apartment" }
      property_description: { type: string, example: "A spacious 3-bedroom apartment." }
      type_id: { type: integer, example: 2 }
      transaction_type: { type: string, example: "sale" }
      price: { type: number, example: 350000 }
      area: { type: number, example: 120 }
      property_address: { type: string, example: "123 Main Street", description: "Должен быть корректным, непустым адрес. Координаты (latitude, longitude) вычисляются с помощью Geoapify Geocoding API при создании/обновлении." }
      city: { type: string, example: "New York" }
      property_status: { type: string, example: "available" }
  propertyhandler.ListPropertiesResponseDoc:
    type: object
    properties:
      properties:
        type: array
        items:
          $ref: '#/definitions/propertyhandler.PropertyDTODoc'
      total: { type: integer, example: 42 }

  # Image DTOs removed
  imagedto.ImageFileDoc:
    type: object
    properties:
      filename: { type: string, example: "photo1.jpg" }
      data: { type: string, example: "base64encodedstring==", description: "Base64 encoded binary data of the image" }
...
  ErrorResponse:
    type: object
    properties:
      code: { type: integer, example: 400 }
      message: { type: string, example: "Некорректные параметры запроса" }
      details:
        type: object
        additionalProperties:
          type: string
  GeocodingErrorResponse:
    type: object
    properties:
      error: { type: string, example: "ошибка геокодирования: external service error" }
      provider: { type: string, example: "geoapify" }
      code: { type: string, example: "request" }
      details: { type: string, example: "timeout while connecting to geo service" }
      address: { type: string, example: "123 Main Street" }
      status_code: { type: integer, example: 504, nullable: true }
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header